+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
  - Add CHANGELOG.md
  - Flows to receive emails from mail server
+ - CLI job to display topic switches